fix(server): Exposed 0.50 compile fixes (less import, limit/offset arity, Short defaults, in-memory sum)
Static review surfaced four Exposed 0.50 API mismatches that would
prevent the rewritten backend from compiling on a fresh JDK 21 build.
All fixed in this commit; no behavioural change.

1. OtpService.kt — 'less' is a member of SqlExpressionBuilder, not a
   top-level function. Fixed import from
       org.jetbrains.exposed.sql.less
   to
       org.jetbrains.exposed.sql.SqlExpressionBuilder.less
   Affects: OtpService.purgeExpired() deleteWhere predicate.

2. AdmissionRouting.kt — Query.limit(Int, Long) is deprecated in
   Exposed 0.50. Switched to the supported chained form:
       .limit(limit).offset(offset)
   Affects: GET /api/v1/admissions/enquiries paginated list.

3. Tables.kt — Column<Short>.default(Int) does not compile because
   Kotlin does not auto-convert Int literals to Short in generic
   type arguments. Fixed by wrapping every literal:
       short(...).default(0.toShort())
       short(...).default(5.toShort())
   Affects: auth_otps {resend_count, attempt_count, max_attempts,
   max_resends} default declarations.

4. UserProfileRouting.kt — Table.select(Expression<*>) for an
   aggregate sum has inconsistent overload resolution across Exposed
   versions. Replaced the SQL-side SUM(size_bytes) aggregate with a
   client-side sumOf {} fold on the rows returned by selectAll(). The
   query is now unambiguous and portable across SQLite (local dev)
   and Postgres (prod). Row counts per school are small enough that
   client-side aggregation has no measurable cost. Removed the now-
   unused 'org.jetbrains.exposed.sql.sum' import.
   Affects: PUT /api/v1/user/profile/gallery storage recompute.

Branch: backend-by-abuzar — extends PR #5.

+++ server/src/main/kotlin/com/littlebridge/vidyaprayag/db/Tables.kt
@@ -70,10 +70,10 @@ object AuthOtpsTable : UUIDTable("auth_otps", "id") {
     val firstSentAt       = timestamp("first_sent_at")
     val expiresAt         = timestamp("expires_at")
 
-    val resendCount       = short("resend_count").default(0)
-    val attemptCount      = short("attempt_count").default(0)
-    val maxAttempts       = short("max_attempts").default(5)
-    val maxResends        = short("max_resends").default(5)
+    val resendCount       = short("resend_count").default(0.toShort())
+    val attemptCount      = short("attempt_count").default(0.toShort())
+    val maxAttempts       = short("max_attempts").default(5.toShort())
+    val maxResends        = short("max_resends").default(5.toShort())
     val resendWindowSecs  = integer("resend_window_secs").default(3600)
 
     val isVerified        = bool("is_verified").default(false)
